removed logging and added error object to operator

+++ model/delete.go
@@ -2,14 +2,17 @@ package model
 
 import (
 	"context"
+	"fmt"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
-	"github.com/rs/zerolog/log"
 	"reflect"
 )
 
 func (o *Operator) Delete(q interface{}) *Operator {
+	if o.Err != nil {
+		return o
+	}
 	payload := reflect.ValueOf(q).Elem()
 	_, err := svc.DeleteItem(context.TODO(), &dynamodb.DeleteItemInput{
 		TableName: aws.String(dynamoDBTableName), Key: map[string]types.AttributeValue{
@@ -18,7 +21,8 @@ func (o *Operator) Delete(q interface{}) *Operator {
 		},
 	})
 	if err != nil {
-		log.Error().Err(err).Msg("failed to delete item")
+		o.Err = fmt.Errorf("encountered an error during Delete operation: %v", err)
+		return o
 	}
 	return o
 }
